PDB: Fix the size of an integer
Fixes an issue from previous CL.

BUG=
R=manzagop@chromium.org

Review URL: https://codereview.chromium.org/1346613003.

+++ syzygy/pdb/pdb_stream_record_unittest.cc
@@ -76,7 +76,7 @@ TEST_F(PdbStreamRecordTest, ReadLeafNumericConstantDirect) {
 }
 
 TEST_F(PdbStreamRecordTest, ReadLeafNumericConstantChar) {
-  const int16_t kVal8 = -42;
+  const int8_t kVal8 = -42;
   const uint16_t kLfChar = Microsoft_Cci_Pdb::LF_CHAR;
   NumericConstant numeric;
